refactor: move tremor components to a separate directory
