+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { GeistSans } from "geist/font/sans";
 import "./globals.css";
-import Sidebar from "@/components/Sidebar";
+import Sidebar from "@/components/tremor/Sidebar";
 
 export const metadata: Metadata = {
   title: "G-Pac",
